+++ R/mcmc-tidiers.R
@@ -1,8 +1,6 @@
 #' Tidying methods for MCMC (Stan, JAGS, etc.) fits
 #' 
-#' MCMC tidiers will soon be deprecated in `broom` and there is no ongoing
-#' development of these functions at this time. MCMC tidiers are being
-#' developed in the `broom.mixed` package, which is not yet on CRAN.
+#' MCMC tidiers are deprecated.
 #'
 #' @param x an object of class \sQuote{"stanfit"}
 #' @param pars (character) specification of which parameters to include
@@ -18,49 +16,6 @@
 #'
 #' @name mcmc_tidiers
 #'
-#' @examples
-#'
-#' \dontrun{
-#'
-#' # Using example from "RStan Getting Started"
-#' # https://github.com/stan-dev/rstan/wiki/RStan-Getting-Started
-#'
-#' model_file <- system.file("extdata", "8schools.stan", package = "broom")
-#'
-#' schools_dat <- list(J = 8,
-#'                     y = c(28,  8, -3,  7, -1,  1, 18, 12),
-#'                     sigma = c(15, 10, 16, 11,  9, 11, 10, 18))
-#'
-#' if (requireNamespace("rstan", quietly = TRUE)) {
-#'   set.seed(2015)
-#'   rstan_example <- stan(file = model_file, data = schools_dat,
-#'                         iter = 100, chains = 2)
-#' }
-#'
-#' }
-#'
-#' if (requireNamespace("rstan", quietly = TRUE)) {
-#'   # the object from the above code was saved as rstan_example.rda
-#'   infile <- system.file("extdata", "rstan_example.rda", package = "broom")
-#'   load(infile)
-#'
-#'   tidy(rstan_example)
-#'   tidy(rstan_example, conf.int = TRUE, pars = "theta")
-#'
-#'   td_mean <- tidy(rstan_example, conf.int = TRUE)
-#'   td_median <- tidy(rstan_example, conf.int = TRUE, estimate.method = "median")
-#'
-#'   library(dplyr)
-#'   library(ggplot2)
-#'   tds <- rbind(mutate(td_mean, method = "mean"),
-#'                mutate(td_median, method = "median"))
-#'
-#'   ggplot(tds, aes(estimate, term)) +
-#'     geom_errorbarh(aes(xmin = conf.low, xmax = conf.high)) +
-#'     geom_point(aes(color = method))
-#' }
-#'
-#'
 #' @export
 tidyMCMC <- function(x,
                      pars, ## ?? do other
@@ -72,6 +27,7 @@ tidyMCMC <- function(x,
                      rhat = FALSE,
                      ess = FALSE,
                      ...) {
+  .Deprecated()
   stan <- inherits(x, "stanfit")
   ss <- if (stan) as.matrix(x, pars = pars) else as.matrix(x)
   ss <- ss[, !colnames(ss) %in% droppars, drop = FALSE] ## drop log-probability info
@@ -125,6 +81,7 @@ tidy.rjags <- function(x,
                        conf.level = 0.95,
                        conf.method = "quantile",
                        ...) {
+  .Deprecated()
   tidyMCMC(coda::as.mcmc(x$BUGS),
     pars,
     estimate.method, conf.int, conf.level,
@@ -135,4 +92,7 @@ tidy.rjags <- function(x,
 
 ##' @rdname mcmc_tidiers
 ##' @export
-tidy.stanfit <- tidyMCMC
+tidy.stanfit <- function(...) {
+  .Deprecated()
+  tidyMCMC(...)
+}